skyglass-operational-console/backend/requirements.txt
Flask==2.2.2
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
python-dotenv==0.21.1
requests==2.28.2
APScheduler==3.10.1
pandas==1.5.3
Flask-Cors==3.0.10
